requirements.txt
click
Mako
pyYaml